IOS-5832 Work with BannerPromotionNotificationFactory
IOS-5832 Add Localizable

IOS-5832 Update

IOS-5832 Updates

IOS-5832 Selfreview
