+++ Tangem/App/Services/ExpressPromotionService/BannerPromotionNotificationFactory.swift
@@ -1,0 +1,96 @@
+//
+//  BannerPromotionNotificationFactory.swift
+//  Tangem
+//
+//  Created by Sergey Balashov on 07.02.2024.
+//  Copyright © 2024 Tangem AG. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+struct BannerPromotionNotificationFactory {
+    func buildMainBannerNotificationInput(
+        promotion: ActivePromotionInfo,
+        dismissAction: @escaping NotificationView.NotificationAction
+    ) -> NotificationViewInput {
+        let event = event(for: promotion, place: .main)
+        return NotificationViewInput(
+            style: .plain,
+            severity: .info,
+            settings: .init(event: event, dismissAction: dismissAction)
+        )
+    }
+
+    func buildTokenBannerNotificationInput(
+        promotion: ActivePromotionInfo,
+        buttonAction: @escaping NotificationView.NotificationButtonTapAction,
+        dismissAction: @escaping NotificationView.NotificationAction
+    ) -> NotificationViewInput {
+        let event = event(for: promotion, place: .tokenDetails)
+        return NotificationViewInput(
+            style: .withButtons([.init(action: buttonAction, actionType: .exchange, isWithLoader: false)]),
+            severity: .info,
+            settings: .init(event: event, dismissAction: dismissAction)
+        )
+    }
+
+    private func event(for promotion: ActivePromotionInfo, place: BannerPromotionPlace) -> BannerNotificationEvent {
+        switch promotion.bannerPromotion {
+        case .changelly:
+            return .changelly(
+                title: changellyTitle(),
+                description: changellyDescription(promotion: promotion, place: place)
+            )
+        }
+    }
+
+    private func changellyTitle() -> NotificationView.Title {
+        let percent = changellyZeroPercent()
+        let attributed = NSMutableAttributedString(
+            string: Localization.mainSwapChangellyPromotionTitle(percent),
+            attributes: [.font: UIFonts.Bold.footnote, .foregroundColor: UIColor(Colors.Text.constantWhite)]
+        )
+
+        if let range = attributed.string.range(of: percent) {
+            let yellow = UIColor(red: 233, green: 253, blue: 2, alpha: 1)
+
+            attributed.addAttribute(
+                .foregroundColor,
+                value: yellow,
+                range: NSRange(range.lowerBound..., in: attributed.string)
+            )
+        }
+
+        return .attributed(attributed)
+    }
+
+    func changellyDescription(promotion: ActivePromotionInfo, place: BannerPromotionPlace) -> String? {
+        let percent = changellyZeroPercent()
+        let formatter = DateFormatter()
+        formatter.dateFormat = "dd"
+
+        switch place {
+        case .main:
+            return Localization.mainSwapChangellyPromotionMessage(
+                percent,
+                formatter.string(from: promotion.timeline.start),
+                formatter.string(from: promotion.timeline.end)
+            )
+        case .tokenDetails:
+            return Localization.tokenSwapChangellyPromotionMessage(
+                percent,
+                formatter.string(from: promotion.timeline.start),
+                formatter.string(from: promotion.timeline.end)
+            )
+        }
+    }
+
+    func changellyZeroPercent() -> String {
+        PercentFormatter().expressRatePercentFormat(
+            value: 0,
+            maximumFractionDigits: 0,
+            minimumFractionDigits: 0
+        )
+    }
+}
