+++ Tangem/Common/Storage/StorageType.swift
@@ -34,7 +34,6 @@ enum StorageType: String {
     case shouldHidingSensitiveInformationSheetShowing = "should_hiding_sensitive_information_sheet_showing"
     case appTheme = "app_theme"
     case userDidSwipeWalletsOnMainScreen = "user_did_swipe_wallets_on_main_screen"
-    case tangemExpressMainPromotionDismissed = "tangem_express_main_promotion_dismissed"
     case mainPromotionDismissed = "main_promotion_dismissed"
     case tokenPromotionDismissed = "token_promotion_dismissed"
 }
